+++ src/sodium_ctx.rs
@@ -32,34 +32,34 @@ impl SodiumCtx {
 
     /// Create a new constant value [`Cell`] in this context.
     pub fn new_cell<A: Clone + Send + 'static>(&self, a: A) -> Cell<A> {
-        Cell::new(self, a)
+        Cell::new_in(self, a)
     }
 
     /// Create a new stream that will never fire in this context.
     pub fn new_stream<A: Clone + Send + 'static>(&self) -> Stream<A> {
-        Stream::new(self)
+        Stream::new_in(self)
     }
 
     /// Create a new [`CellSink`] for interfacing I/O and FRP.
     pub fn new_cell_sink<A: Clone + Send + 'static>(&self, a: A) -> CellSink<A> {
-        CellSink::new(self, a)
+        CellSink::new_in(self, a)
     }
 
     /// Create a new [`StreamSink`] for interfacing I/O and FRP.
     pub fn new_stream_sink<A: Clone + Send + 'static>(&self) -> StreamSink<A> {
-        StreamSink::new(self)
+        StreamSink::new_in(self)
     }
 
     /// Create a new [`CellLoop`] to act as a forward reference for a
     /// [`Cell`] that will be created later.
     pub fn new_cell_loop<A: Clone + Send + 'static>(&self) -> CellLoop<A> {
-        CellLoop::new(self)
+        CellLoop::new_in(self)
     }
 
     /// Create a new [`StreamLoop`] to act as a forward reference for
     /// a [`Stream`] that will be created later.
     pub fn new_stream_loop<A: Clone + Send + 'static>(&self) -> StreamLoop<A> {
-        StreamLoop::new(self)
+        StreamLoop::new_in(self)
     }
 
     /// Create a new [`StreamSink`] with a combining function that
@@ -72,7 +72,7 @@ impl SodiumCtx {
         &self,
         coalescer: COALESCER,
     ) -> StreamSink<A> {
-        StreamSink::new_with_coalescer(self, coalescer)
+        StreamSink::new_with_coalescer_in(self, coalescer)
     }
 
     /// Run the given function inside a single Sodium transaction,
@@ -87,7 +87,7 @@ impl SodiumCtx {
     /// the returned [`Transaction`] is dropped or
     /// [Transaction::close] is called explicitly.
     pub fn new_transaction(&self) -> Transaction {
-        Transaction::new(self)
+        Transaction::new_in(self)
     }
 
     /// Execute the given code after the current transaction is
@@ -106,6 +106,6 @@ impl SodiumCtx {
         A: Clone + Send + 'static,
         K: Send + Sync + Eq + Hash + 'static,
     {
-        Router::new(self, in_stream, selector)
+        Router::new_in(self, in_stream, selector)
     }
 }